backend/requirements.txt
# Web Frameworks
fastapi[standard]

# Machine Learning
numpy
torch

# Database
psycopg2
sqlalchemy
sqlalchemy-utils

# Auth & Security
bcrypt
pyjwt

# Validation & Settings
pydantic
pydantic[email]
python-dotenv